pyproject.toml
[project]
name = "wakeabcbot"
version = "0.1.0"
description = "A telegram bot that helps you stay informed about WakeABC.com inventory"
authors = [
    {name = "Brandon Squizzato",email = "bsquizzato@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "schedule>=1.2.2",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.12.14"
]

[project.scripts]
wakeabcbot = "wakeabcbot.main:main"

[tool.poetry]
packages = [{include = "wakeabcbot", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
